+++ lib/src/pod_list_builder.dart
@@ -42,7 +42,7 @@ class PodListBuilder<T extends dynamic> extends StatefulWidget {
   /// A builder function that takes the current `BuildContext`, the child `Widget`
   /// and a list of current values of the `pods`. It returns a `Widget` that is
   /// rebuilt every time one of the `Pod` objects notifies its listeners.
-  final Widget Function(BuildContext, Widget?, List<T?>) builder;
+  final Widget? Function(BuildContext, Widget?, List<T?>) builder;
 
   //
   //
@@ -55,7 +55,7 @@ class PodListBuilder<T extends dynamic> extends StatefulWidget {
   /// - `pods`: A list of `Pod` objects to listen to.
   /// - `builder`: A function that builds the UI based on the `pods`.
   /// - `child` (optional): A child widget to be passed to the builder function.
-  const PodListBuilder({
+  const PodListBuilder.def({
     super.key,
     this.pods = const [],
     required this.builder,
@@ -72,12 +72,12 @@ class PodListBuilder<T extends dynamic> extends StatefulWidget {
   /// - `key`: An identifier for this widget in the widget tree.
   /// - `pods`: A list of `Pod` objects to listen to.
   /// - `builder`: A function that builds the UI based on the `pods`.
-  factory PodListBuilder.values({
+  factory PodListBuilder({
     Key? key,
     List<Pod<T>> pods = const [],
-    required Widget Function(List<T?>) builder,
+    required Widget? Function(List<T?>) builder,
   }) {
-    return PodListBuilder<T>(
+    return PodListBuilder<T>.def(
       pods: pods,
       builder: (_, __, values) => builder(values),
     );
@@ -114,7 +114,6 @@ class PodListBuilderState<T extends dynamic> extends State<PodListBuilder<T>> {
   @override
   void didUpdateWidget(PodListBuilder<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
-    // Remove listeners from old Pods and add listeners to new Pods.
     for (final oldPod in oldWidget.pods) {
       oldPod?.removeListener(_update);
     }
@@ -140,12 +139,12 @@ class PodListBuilderState<T extends dynamic> extends State<PodListBuilder<T>> {
 
   @override
   Widget build(BuildContext context) {
-    // Build the widget using the provided builder function.
     return widget.builder(
-      context,
-      widget.child,
-      widget.pods.map((e) => e?.value).toList(),
-    );
+          context,
+          widget.child,
+          widget.pods.map((e) => e?.value).toList(),
+        ) ??
+        const SizedBox.shrink();
   }
 
   //
@@ -157,7 +156,7 @@ class PodListBuilderState<T extends dynamic> extends State<PodListBuilder<T>> {
     // Remove listeners from all Pods when the widget is disposed.
     for (final pod in widget.pods) {
       pod?.removeListener(_update);
-      pod?.disposeIfTemp();
+      pod?.disposeIfMarkedAsTemp();
     }
     super.dispose();
   }